test: split config flow user tests incrementally
